requirements.txt
line-bot-sdk
flask
pyquery
selenium

aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
DateTime==5.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
Flask==2.2.2
frozenlist==1.3.3
future==0.18.3
greenlet==2.0.1
h11==0.14.0
html5lib==1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.1
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.1
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
webencodings==0.5.1
Werkzeug==2.2.2
wsproto==1.2.0
XlsxWriter==3.0.7
yarl==1.8.2
zope.interface==5.5.2
